+++ src/rospack.cpp
@@ -27,7 +27,7 @@
 
 #include "rospack/rospack.h"
 #include "utils.h"
-#include "tinyxml.h"
+#include "tinyxml2.h"
 
 #include <boost/algorithm/string.hpp>
 #include <boost/filesystem.hpp>
@@ -81,6 +81,8 @@
 #define PyUnicode_FromString PyString_FromString
 #endif
 
+using namespace tinyxml2;
+
 // TODO:
 //   recrawl on:
 //     package not found in cache
@@ -115,7 +117,7 @@ static const int MAX_CRAWL_DEPTH = 1000;
 static const int MAX_DEPENDENCY_DEPTH = 1000;
 static const double DEFAULT_MAX_CACHE_AGE = 60.0;
 
-TiXmlElement* get_manifest_root(Stackage* stackage);
+XMLElement* get_manifest_root(Stackage* stackage);
 double time_since_epoch();
 
 #ifdef __APPLE__
@@ -147,10 +149,12 @@ class Stackage
     std::string manifest_path_;
     // \brief filename of the stackage manifest
     std::string manifest_name_;
+    // \brief package's license with a support for multi-license.
+    std::vector<std::string> licenses_;
     // \brief have we already loaded the manifest?
     bool manifest_loaded_;
     // \brief TinyXML structure, filled in during parsing
-    TiXmlDocument manifest_;
+    XMLDocument manifest_;
     std::vector<Stackage*> deps_;
     bool deps_computed_;
     bool is_wet_package_;
@@ -176,14 +180,20 @@ class Stackage
       assert(is_wet_package_);
       assert(manifest_loaded_);
       // get name from package.xml instead of folder name
-      TiXmlElement* root = get_manifest_root(this);
-      for(TiXmlElement* el = root->FirstChildElement("name"); el; el = el->NextSiblingElement("name"))
+      XMLElement* root = get_manifest_root(this);
+      for(XMLElement* el = root->FirstChildElement("name"); el; el = el->NextSiblingElement("name"))
       {
         name_ = el->GetText();
         break;
       }
+      // Get license texts, where there may be multiple elements for.
+      std::string tagname_license = "license";
+      for(XMLElement* el = root->FirstChildElement(tagname_license.c_str()); el; el = el->NextSiblingElement(tagname_license.c_str()))
+      {
+        licenses_.push_back(el->GetText());
+      }
       // check if package is a metapackage
-      for(TiXmlElement* el = root->FirstChildElement("export"); el; el = el->NextSiblingElement("export"))
+      for(XMLElement* el = root->FirstChildElement("export"); el; el = el->NextSiblingElement("export"))
       {
         if(el->FirstChildElement("metapackage"))
         {
@@ -722,8 +732,8 @@ Rosstackage::rosdeps(const std::string& name, bool direct,
 void
 Rosstackage::_rosdeps(Stackage* stackage, std::set<std::string>& rosdeps, const char* tag_name)
 {
-  TiXmlElement* root = get_manifest_root(stackage);
-  for(TiXmlElement* ele = root->FirstChildElement(tag_name);
+  XMLElement* root = get_manifest_root(stackage);
+  for(XMLElement* ele = root->FirstChildElement(tag_name);
       ele;
       ele = ele->NextSiblingElement(tag_name))
   {
@@ -765,8 +775,8 @@ Rosstackage::vcs(const std::string& name, bool direct,
         it != deps_vec.end();
         ++it)
     {
-      TiXmlElement* root = get_manifest_root(*it);
-      for(TiXmlElement* ele = root->FirstChildElement(MANIFEST_TAG_VERSIONCONTROL);
+      XMLElement* root = get_manifest_root(*it);
+      for(XMLElement* ele = root->FirstChildElement(MANIFEST_TAG_VERSIONCONTROL);
           ele;
           ele = ele->NextSiblingElement(MANIFEST_TAG_VERSIONCONTROL))
       {
@@ -1007,16 +1017,16 @@ Rosstackage::exports_dry_package(Stackage* stackage, const std::string& lang,
                      const std::string& attrib,
                      std::vector<std::string>& flags)
 {
-  TiXmlElement* root = get_manifest_root(stackage);
-  for(TiXmlElement* ele = root->FirstChildElement(MANIFEST_TAG_EXPORT);
+  XMLElement* root = get_manifest_root(stackage);
+  for(XMLElement* ele = root->FirstChildElement(MANIFEST_TAG_EXPORT);
       ele;
       ele = ele->NextSiblingElement(MANIFEST_TAG_EXPORT))
   {
     bool os_match = false;
     const char *best_match = NULL;
-    for(TiXmlElement* ele2 = ele->FirstChildElement(lang);
+    for(XMLElement* ele2 = ele->FirstChildElement(lang.c_str());
         ele2;
-        ele2 = ele2->NextSiblingElement(lang))
+        ele2 = ele2->NextSiblingElement(lang.c_str()))
     {
       const char *os_str;
       if ((os_str = ele2->Attribute("os")))
@@ -1113,14 +1123,14 @@ Rosstackage::plugins(const std::string& name, const std::string& attrib,
       it != stackages.end();
       ++it)
   {
-    TiXmlElement* root = get_manifest_root(*it);
-    for(TiXmlElement* ele = root->FirstChildElement(MANIFEST_TAG_EXPORT);
+    XMLElement* root = get_manifest_root(*it);
+    for(XMLElement* ele = root->FirstChildElement(MANIFEST_TAG_EXPORT);
         ele;
         ele = ele->NextSiblingElement(MANIFEST_TAG_EXPORT))
     {
-      for(TiXmlElement* ele2 = ele->FirstChildElement(name);
+      for(XMLElement* ele2 = ele->FirstChildElement(name.c_str());
           ele2;
-          ele2 = ele2->NextSiblingElement(name))
+          ele2 = ele2->NextSiblingElement(name.c_str()))
       {
         const char *att_str;
         if((att_str = ele2->Attribute(attrib.c_str())))
@@ -1544,7 +1554,7 @@ Rosstackage::loadManifest(Stackage* stackage)
   if(stackage->manifest_loaded_)
     return;
 
-  if(!stackage->manifest_.LoadFile(stackage->manifest_path_))
+  if(stackage->manifest_.LoadFile(stackage->manifest_path_.c_str()) != XML_SUCCESS)
   {
     std::string errmsg = std::string("error parsing manifest of package ") +
             stackage->name_ + " at " + stackage->manifest_path_;
@@ -1590,14 +1600,14 @@ Rosstackage::computeDeps(Stackage* stackage, bool ignore_errors, bool ignore_mis
 void
 Rosstackage::computeDepsInternal(Stackage* stackage, bool ignore_errors, const std::string& depend_tag, bool ignore_missing)
 {
-  TiXmlElement* root;
+  XMLElement* root;
   root = get_manifest_root(stackage);
 
-  TiXmlNode *dep_node = NULL;
   const char* dep_pkgname;
-  while((dep_node = root->IterateChildren(depend_tag, dep_node)))
+  for(XMLElement *dep_ele = root->FirstChildElement(depend_tag.c_str());
+      dep_ele;
+      dep_ele = dep_ele->NextSiblingElement(depend_tag.c_str()))
   {
-    TiXmlElement *dep_ele = dep_node->ToElement();
     if (!stackage->is_wet_package_)
     {
       dep_pkgname = dep_ele->Attribute(tag_.c_str());
@@ -1994,11 +2004,12 @@ Rosstackage::writeCache()
   }
   else
   {
-    char tmp_cache_dir[PATH_MAX];
-    char tmp_cache_path[PATH_MAX];
-    strncpy(tmp_cache_dir, cache_path.c_str(), sizeof(tmp_cache_dir));
+    size_t len = cache_path.size() + 1;
+    char *tmp_cache_dir = new char[len];
+    strncpy(tmp_cache_dir, cache_path.c_str(), len);
 #if defined(_MSC_VER)
     // No dirname on Windows; use _splitpath_s instead
+    char tmp_cache_path[PATH_MAX];
     char drive[_MAX_DRIVE], dir[_MAX_DIR], fname[_MAX_FNAME], ext[_MAX_EXT];
     _splitpath_s(tmp_cache_dir, drive, _MAX_DRIVE, dir, _MAX_DIR, fname, _MAX_FNAME,
                  ext, _MAX_EXT);
@@ -2006,11 +2017,15 @@ Rosstackage::writeCache()
     _makepath_s(full_dir, _MAX_DRIVE + _MAX_DIR, drive, dir, NULL, NULL);
     snprintf(tmp_cache_path, sizeof(tmp_cache_path), "%s\\.rospack_cache.XXXXXX", full_dir);
 #elif defined(__MINGW32__)
+    char tmp_cache_path[PATH_MAX];
     char* temp_name = tempnam(dirname(tmp_cache_dir),".rospack_cache.");
     snprintf(tmp_cache_path, sizeof(tmp_cache_path), temp_name);
     delete temp_name;
 #else
-    snprintf(tmp_cache_path, sizeof(tmp_cache_path), "%s/.rospack_cache.XXXXXX", dirname(tmp_cache_dir));
+    char *temp_dirname = dirname(tmp_cache_dir);
+    len = strlen(temp_dirname) + 22 + 1;
+    char *tmp_cache_path = new char[len];
+    snprintf(tmp_cache_path, len, "%s/.rospack_cache.XXXXXX", temp_dirname);
 #endif
 #if defined(__MINGW32__)
     // There is no equivalent of mkstemp or _mktemp_s on mingw, so we resort to a slightly problematic
@@ -2068,6 +2083,10 @@ Rosstackage::writeCache()
         }
       }
     }
+    delete[] tmp_cache_dir;
+#if !defined(_MSC_VER) && !defined(__MINGW32__)
+    delete[] tmp_cache_path;
+#endif
   }
 }
 
@@ -2310,10 +2329,10 @@ std::string Rosstack::get_manifest_type()
   return "stack";
 }
 
-TiXmlElement*
+XMLElement*
 get_manifest_root(Stackage* stackage)
 {
-  TiXmlElement* ele = stackage->manifest_.RootElement();
+  XMLElement* ele = stackage->manifest_.RootElement();
   if(!ele)
   {
     std::string errmsg = std::string("error parsing manifest of package ") +