+++ CHANGELOG.rst
@@ -2,16 +2,29 @@
 Changelog for package rospack
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-2.2.8 (2017-03-06)
+2.4.2 (2017-07-27)
+------------------
+
+2.4.1 (2017-02-27)
+------------------
+* fix inverted result code check, regression from 2.4.0 (`#70 <https://github.com/ros/rospack/issues/70>`_)
+
+2.4.0 (2017-02-22)
 ------------------
 * make some deps* functions public (`#65 <https://github.com/ros/rospack/pull/65>`_)
+* switch from TinyXML to TinyXML2 (`#62 <https://github.com/ros/rospack/pull/62>`_)
 
-2.2.7 (2016-06-27)
+2.3.2 (2017-02-14)
+------------------
+* add license field in Stackage class (`#66 <https://github.com/ros/rospack/issues/66>`_)
+
+2.3.1 (2016-09-02)
+------------------
+* fix FTBFS on hurd-i386 (`#64 <https://github.com/ros/rospack/issues/64>`_)
+
+2.3.0 (2016-03-09)
 ------------------
 * allow caching of rospack results (`#49 <https://github.com/ros/rospack/issues/49>`_)
-
-2.2.6 (2016-03-09)
-------------------
 * fix memory leak in Rosstackage::addStackage (`#59 <https://github.com/ros/rospack/issues/59>`_)
 * return false in depsOnDetail if the package name in rospack plugins can not be found (`#51 <https://github.com/ros/rospack/issues/51>`_)
 * #undef symbols before #defining them to avoid preprocessor warnings in the case that they were already #defined (`#50 <https://github.com/ros/rospack/issues/50>`_)